chore: moved dto to lower level module
M  src/main/java/com/tuvarna/phd/controller/AuthController.java
M  src/main/java/com/tuvarna/phd/controller/CandidateController.java
M  src/main/java/com/tuvarna/phd/controller/CommitteeController.java
M  src/main/java/com/tuvarna/phd/controller/DoctoralCenterAdminController.java
M  src/main/java/com/tuvarna/phd/controller/DoctoralCenterController.java
M  src/main/java/com/tuvarna/phd/controller/FileController.java
M  src/main/java/com/tuvarna/phd/controller/LogController.java
M  src/main/java/com/tuvarna/phd/controller/NotificationController.java
M  src/main/java/com/tuvarna/phd/controller/PhdController.java
R  src/main/java/com/tuvarna/phd/service/dto/BlobDataDTO.java -> src/main/java/com/tuvarna/phd/dto/BlobDataDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/CandidateDTO.java -> src/main/java/com/tuvarna/phd/dto/CandidateDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/CommitteeDTO.java -> src/main/java/com/tuvarna/phd/dto/CommitteeDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/DoctoralCenterDTO.java -> src/main/java/com/tuvarna/phd/dto/DoctoralCenterDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/FileBlobDTO.java -> src/main/java/com/tuvarna/phd/dto/FileBlobDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/FilenameDTO.java -> src/main/java/com/tuvarna/phd/dto/FilenameDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/IdDTO.java -> src/main/java/com/tuvarna/phd/dto/IdDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/LogDTO.java -> src/main/java/com/tuvarna/phd/dto/LogDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/NotificationClientDTO.java -> src/main/java/com/tuvarna/phd/dto/NotificationClientDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/NotificationDTO.java -> src/main/java/com/tuvarna/phd/dto/NotificationDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/PhdDTO.java -> src/main/java/com/tuvarna/phd/dto/PhdDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/RoleDTO.java -> src/main/java/com/tuvarna/phd/dto/RoleDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/UnauthorizedUsersDTO.java -> src/main/java/com/tuvarna/phd/dto/UnauthorizedUsersDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/UserDTO.java -> src/main/java/com/tuvarna/phd/dto/UserDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/UserPrincipalDTO.java -> src/main/java/com/tuvarna/phd/dto/UserPrincipalDTO.java
R  src/main/java/com/tuvarna/phd/service/dto/sendLogDTO.java -> src/main/java/com/tuvarna/phd/dto/sendLogDTO.java
M  src/main/java/com/tuvarna/phd/mapper/CandidateMapper.java
M  src/main/java/com/tuvarna/phd/mapper/LogMapper.java
M  src/main/java/com/tuvarna/phd/mapper/NotificationMapper.java
M  src/main/java/com/tuvarna/phd/mapper/PhdMapper.java
M  src/main/java/com/tuvarna/phd/mapper/UnauthorizedUsersMapper.java
M  src/main/java/com/tuvarna/phd/service/AuthService.java
M  src/main/java/com/tuvarna/phd/service/AuthServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/CandidateService.java
M  src/main/java/com/tuvarna/phd/service/CandidateServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/DoctoralCenterService.java
M  src/main/java/com/tuvarna/phd/service/DoctoralCenterServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/LogService.java
M  src/main/java/com/tuvarna/phd/service/LogServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/NotificationService.java
M  src/main/java/com/tuvarna/phd/service/NotificationServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/S3ClientService.java
M  src/main/java/com/tuvarna/phd/service/S3ClientServiceImpl.java
M  src/main/java/com/tuvarna/phd/validator/DoctoralCenterValidator.java
M  src/main/java/com/tuvarna/phd/validator/NotificationValidator.java
