+++ src/main/java/com/tuvarna/phd/controller/LogController.java
@@ -3,9 +3,9 @@ package com.tuvarna.phd.controller;
 import com.tuvarna.phd.mapper.LogMapper;
 import com.tuvarna.phd.repository.DoctoralCenterRepository;
 import com.tuvarna.phd.service.LogService;
-import com.tuvarna.phd.service.dto.LogDTO;
-import com.tuvarna.phd.service.dto.UserPrincipalDTO;
-import com.tuvarna.phd.service.dto.sendLogDTO;
+import com.tuvarna.phd.dto.LogDTO;
+import com.tuvarna.phd.dto.UserPrincipalDTO;
+import com.tuvarna.phd.dto.sendLogDTO;
 import com.tuvarna.phd.validator.LogValidator;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
