+++ src/main/java/com/tuvarna/phd/dto/CandidateDTO.java
@@ -1,4 +1,4 @@
-package com.tuvarna.phd.service.dto;
+package com.tuvarna.phd.dto;
 
 import jakarta.annotation.Nullable;
 import lombok.AllArgsConstructor;
