+++ src/main/java/com/tuvarna/phd/dto/DoctoralCenterDTO.java
@@ -1,18 +1,12 @@
-package com.tuvarna.phd.service.dto;
+package com.tuvarna.phd.dto;
 
 import io.smallrye.common.constraint.NotNull;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
-@Schema(name = "UserDTO", description = "User DTO")
+@Schema(name = "DoctoralCenterDTO", description = "Doctoral Center DTO")
 @Data
-@AllArgsConstructor
-public class UserDTO {
-
-  @NotNull
-  @Schema(title = "id", required = true)
-  private Long id;
+public class DoctoralCenterDTO {
 
   @NotNull
   @Schema(title = "oid", required = true)
@@ -26,6 +20,11 @@ public class UserDTO {
   @Schema(title = "email", required = true)
   private String email;
 
+  @NotNull
+  @Schema(title = "picture", required = true)
+  private String picture;
+
+  @NotNull
   @Schema(title = "role", required = false)
   private String role;
 }
