+++ src/main/java/com/tuvarna/phd/dto/FilenameDTO.java
@@ -1,4 +1,4 @@
-package com.tuvarna.phd.service.dto;
+package com.tuvarna.phd.dto;
 
 import io.smallrye.common.constraint.NotNull;
 import lombok.AllArgsConstructor;
