+++ src/main/java/com/tuvarna/phd/dto/LogDTO.java
@@ -1,15 +1,21 @@
-package com.tuvarna.phd.service.dto;
+package com.tuvarna.phd.dto;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.smallrye.common.constraint.NotNull;
+import io.smallrye.common.constraint.Nullable;
 import java.time.Instant;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
-@Schema(name = "sendLogDTO", description = "Log DTO")
+@Schema(name = "LogDTO", description = "Log DTO")
 @Data
 @AllArgsConstructor
-public class sendLogDTO {
+public class LogDTO {
+
+  @Nullable
+  @Schema(title = "id", required = true)
+  private String id;
 
   @NotNull
   @Schema(title = "description", required = true)
@@ -26,4 +32,8 @@ public class sendLogDTO {
   @NotNull
   @Schema(title = "level", required = true)
   private String level;
+
+  @Schema(title = "user", required = true)
+  @JsonProperty("user")
+  private UserPrincipalDTO userPrincipalDTO;
 }
