+++ src/main/java/com/tuvarna/phd/dto/UserPrincipalDTO.java
@@ -1,4 +1,4 @@
-package com.tuvarna.phd.service.dto;
+package com.tuvarna.phd.dto;
 
 import io.smallrye.common.constraint.NotNull;
 import jakarta.annotation.Nullable;
