+++ src/main/java/com/tuvarna/phd/service/CandidateService.java
@@ -1,6 +1,6 @@
 package com.tuvarna.phd.service;
 
-import com.tuvarna.phd.service.dto.CandidateDTO;
+import com.tuvarna.phd.dto.CandidateDTO;
 
 public sealed interface CandidateService permits CandidateServiceImpl {
   void register(CandidateDTO candidateDTO);
