+++ src/main/java/com/tuvarna/phd/service/DoctoralCenterService.java
@@ -1,10 +1,10 @@
 package com.tuvarna.phd.service;
 
+import com.tuvarna.phd.dto.CandidateDTO;
+import com.tuvarna.phd.dto.RoleDTO;
+import com.tuvarna.phd.dto.UnauthorizedUsersDTO;
+import com.tuvarna.phd.dto.UserDTO;
 import com.tuvarna.phd.entity.UnauthorizedUsers;
-import com.tuvarna.phd.service.dto.CandidateDTO;
-import com.tuvarna.phd.service.dto.RoleDTO;
-import com.tuvarna.phd.service.dto.UnauthorizedUsersDTO;
-import com.tuvarna.phd.service.dto.UserDTO;
 import io.smallrye.mutiny.Uni;
 import java.util.List;
