+++ src/main/java/com/tuvarna/phd/service/DoctoralCenterServiceImpl.java
@@ -14,10 +14,10 @@ import com.tuvarna.phd.repository.DoctoralCenterRoleRepository;
 import com.tuvarna.phd.repository.PhdRepository;
 import com.tuvarna.phd.repository.PhdStatusRepository;
 import com.tuvarna.phd.repository.UnauthorizedUsersRepository;
-import com.tuvarna.phd.service.dto.CandidateDTO;
-import com.tuvarna.phd.service.dto.RoleDTO;
-import com.tuvarna.phd.service.dto.UnauthorizedUsersDTO;
-import com.tuvarna.phd.service.dto.UserDTO;
+import com.tuvarna.phd.dto.CandidateDTO;
+import com.tuvarna.phd.dto.RoleDTO;
+import com.tuvarna.phd.dto.UnauthorizedUsersDTO;
+import com.tuvarna.phd.dto.UserDTO;
 import io.quarkus.mailer.Mail;
 import io.quarkus.mailer.reactive.ReactiveMailer;
 import io.smallrye.mutiny.Uni;
