+++ src/main/java/com/tuvarna/phd/service/LogService.java
@@ -1,6 +1,6 @@
 package com.tuvarna.phd.service;
 
-import com.tuvarna.phd.service.dto.LogDTO;
+import com.tuvarna.phd.dto.LogDTO;
 import java.util.List;
 
 public sealed interface LogService permits LogServiceImpl {
