+++ src/main/java/com/tuvarna/phd/service/NotificationService.java
@@ -1,8 +1,8 @@
 package com.tuvarna.phd.service;
 
-import com.tuvarna.phd.service.dto.IdDTO;
-import com.tuvarna.phd.service.dto.NotificationClientDTO;
-import com.tuvarna.phd.service.dto.NotificationDTO;
+import com.tuvarna.phd.dto.IdDTO;
+import com.tuvarna.phd.dto.NotificationClientDTO;
+import com.tuvarna.phd.dto.NotificationDTO;
 import java.util.List;
 
 public sealed interface NotificationService permits NotificationServiceImpl {
