+++ src/main/java/com/tuvarna/phd/service/NotificationServiceImpl.java
@@ -5,9 +5,9 @@ import com.tuvarna.phd.exception.NotificationException;
 import com.tuvarna.phd.mapper.NotificationMapper;
 import com.tuvarna.phd.repository.DoctoralCenterRepository;
 import com.tuvarna.phd.repository.NotificationRepository;
-import com.tuvarna.phd.service.dto.IdDTO;
-import com.tuvarna.phd.service.dto.NotificationClientDTO;
-import com.tuvarna.phd.service.dto.NotificationDTO;
+import com.tuvarna.phd.dto.IdDTO;
+import com.tuvarna.phd.dto.NotificationClientDTO;
+import com.tuvarna.phd.dto.NotificationDTO;
 import io.smallrye.mutiny.Uni;
 import io.vertx.mutiny.pgclient.PgPool;
 import io.vertx.mutiny.sqlclient.Tuple;
